Fix react-perf/jsx-*
