+++ crates/oxc_linter/src/rules.rs
@@ -168,10 +168,10 @@ mod react {
 }
 
 mod react_perf {
-    pub mod no_jsx_as_prop;
-    pub mod no_new_array_as_prop;
-    pub mod no_new_function_as_props;
-    pub mod no_new_object_as_prop;
+    pub mod jsx_no_jsx_as_prop;
+    pub mod jsx_no_new_array_as_prop;
+    pub mod jsx_no_new_function_as_props;
+    pub mod jsx_no_new_object_as_prop;
 }
 
 mod unicorn {
@@ -522,10 +522,10 @@ oxc_macros::declare_all_lint_rules! {
     react::no_is_mounted,
     react::no_unknown_property,
     react::require_render_return,
-    react_perf::no_jsx_as_prop,
-    react_perf::no_new_array_as_prop,
-    react_perf::no_new_function_as_props,
-    react_perf::no_new_object_as_prop,
+    react_perf::jsx_no_jsx_as_prop,
+    react_perf::jsx_no_new_array_as_prop,
+    react_perf::jsx_no_new_function_as_props,
+    react_perf::jsx_no_new_object_as_prop,
     import::default,
     import::no_named_as_default_member,
     import::no_named_as_default,
