+++ crates/oxc_linter/src/rules/react_perf/jsx_no_jsx_as_prop.rs
@@ -9,44 +9,38 @@ use oxc_diagnostics::{
 use oxc_macros::declare_oxc_lint;
 use oxc_span::Span;
 
-use crate::{
-    context::LintContext,
-    rule::Rule,
-    utils::{get_prop_value, is_constructor_matching_name},
-    AstNode,
-};
+use crate::{context::LintContext, rule::Rule, utils::get_prop_value, AstNode};
 
 #[derive(Debug, Error, Diagnostic)]
-#[error("eslint-plugin-react-perf(no-new-array-as-prop): JSX attribute values should not contain Arrays created in the same scope.")]
+#[error(
+    "eslint-plugin-react-perf(jsx-no-jsx-as-prop): JSX attribute values should not contain other JSX."
+)]
 #[diagnostic(severity(warning), help(r"simplify props or memoize props in the parent component (https://react.dev/reference/react/memo#my-component-rerenders-when-a-prop-is-an-object-or-array)."))]
-struct NoNewArrayAsPropDiagnostic(#[label] pub Span);
+struct NoJsxAsPropDiagnostic(#[label] pub Span);
 
 #[derive(Debug, Default, Clone)]
-pub struct NoNewArrayAsProp;
+pub struct NoJsxAsProp;
 
 declare_oxc_lint!(
     /// ### What it does
     ///
-    /// Prevent Arrays that are local to the current method from being used as values of JSX props
+    /// Prevent JSX that are local to the current method from being used as values of JSX props
     ///
     /// ### Example
     /// ```javascript
     /// // Bad
-    /// <Item list={[]} />
-
-    /// <Item list={new Array()} />
-    /// <Item list={Array()} />
-    /// <Item list={this.props.list || []} />
-    /// <Item list={this.props.list ? this.props.list : []} />
+    /// <Item jsx={<SubItem />} />
+    /// <Item jsx={this.props.jsx || <SubItem />} />
+    /// <Item jsx={this.props.jsx ? this.props.jsx : <SubItem />} />
     ///
     /// // Good
-    /// <Item list={this.props.list} />
+    /// <Item callback={this.props.jsx} />
     /// ```
-    NoNewArrayAsProp,
+    NoJsxAsProp,
     correctness
 );
 
-impl Rule for NoNewArrayAsProp {
+impl Rule for NoJsxAsProp {
     fn run<'a>(&self, node: &AstNode<'a>, ctx: &LintContext<'a>) {
         if let AstKind::JSXElement(jsx_elem) = node.kind() {
             check_jsx_element(jsx_elem, ctx);
@@ -63,7 +57,7 @@ fn check_jsx_element<'a>(jsx_elem: &JSXElement<'a>, ctx: &LintContext<'a>) {
                 ..
             })) => {
                 if let Some(span) = check_expression(expr) {
-                    ctx.diagnostic(NoNewArrayAsPropDiagnostic(span));
+                    ctx.diagnostic(NoJsxAsPropDiagnostic(span));
                 }
             }
             _ => {}
@@ -73,21 +67,7 @@ fn check_jsx_element<'a>(jsx_elem: &JSXElement<'a>, ctx: &LintContext<'a>) {
 
 fn check_expression(expr: &Expression) -> Option<Span> {
     match expr.without_parenthesized() {
-        Expression::ArrayExpression(expr) => Some(expr.span),
-        Expression::CallExpression(expr) => {
-            if is_constructor_matching_name(&expr.callee, "Array") {
-                Some(expr.span)
-            } else {
-                None
-            }
-        }
-        Expression::NewExpression(expr) => {
-            if is_constructor_matching_name(&expr.callee, "Array") {
-                Some(expr.span)
-            } else {
-                None
-            }
-        }
+        Expression::JSXElement(expr) => Some(expr.span),
         Expression::LogicalExpression(expr) => {
             check_expression(&expr.left).or_else(|| check_expression(&expr.right))
         }
@@ -102,18 +82,14 @@ fn check_expression(expr: &Expression) -> Option<Span> {
 fn test() {
     use crate::tester::Tester;
 
-    let pass = vec![r"<Item list={this.props.list} />"];
+    let pass = vec![r"<Item callback={this.props.jsx} />"];
 
     let fail = vec![
-        r"<Item list={[]} />",
-        r"<Item list={new Array()} />",
-        r"<Item list={Array()} />",
-        r"<Item list={this.props.list || []} />",
-        r"<Item list={this.props.list ? this.props.list : []} />",
-        r"<Item list={this.props.list || (this.props.arr ? this.props.arr : [])} />",
+        r"<Item jsx={<SubItem />} />",
+        r"<Item jsx={this.props.jsx || <SubItem />} />",
+        r"<Item jsx={this.props.jsx ? this.props.jsx : <SubItem />} />",
+        r"<Item jsx={this.props.jsx || (this.props.component ? this.props.component : <SubItem />)} />",
     ];
 
-    Tester::new(NoNewArrayAsProp::NAME, pass, fail)
-        .with_react_perf_plugin(true)
-        .test_and_snapshot();
+    Tester::new(NoJsxAsProp::NAME, pass, fail).with_react_perf_plugin(true).test_and_snapshot();
 }
