+++ crates/oxc_linter/src/rules/react_perf/jsx_no_new_array_as_prop.rs
@@ -9,38 +9,44 @@ use oxc_diagnostics::{
 use oxc_macros::declare_oxc_lint;
 use oxc_span::Span;
 
-use crate::{context::LintContext, rule::Rule, utils::get_prop_value, AstNode};
+use crate::{
+    context::LintContext,
+    rule::Rule,
+    utils::{get_prop_value, is_constructor_matching_name},
+    AstNode,
+};
 
 #[derive(Debug, Error, Diagnostic)]
-#[error(
-    "eslint-plugin-react-perf(no-jsx-as-prop): JSX attribute values should not contain other JSX."
-)]
+#[error("eslint-plugin-react-perf(jsx-no-new-array-as-prop): JSX attribute values should not contain Arrays created in the same scope.")]
 #[diagnostic(severity(warning), help(r"simplify props or memoize props in the parent component (https://react.dev/reference/react/memo#my-component-rerenders-when-a-prop-is-an-object-or-array)."))]
-struct NoJsxAsPropDiagnostic(#[label] pub Span);
+struct NoNewArrayAsPropDiagnostic(#[label] pub Span);
 
 #[derive(Debug, Default, Clone)]
-pub struct NoJsxAsProp;
+pub struct NoNewArrayAsProp;
 
 declare_oxc_lint!(
     /// ### What it does
     ///
-    /// Prevent JSX that are local to the current method from being used as values of JSX props
+    /// Prevent Arrays that are local to the current method from being used as values of JSX props
     ///
     /// ### Example
     /// ```javascript
     /// // Bad
-    /// <Item jsx={<SubItem />} />
-    /// <Item jsx={this.props.jsx || <SubItem />} />
-    /// <Item jsx={this.props.jsx ? this.props.jsx : <SubItem />} />
+    /// <Item list={[]} />
+
+    /// <Item list={new Array()} />
+    /// <Item list={Array()} />
+    /// <Item list={this.props.list || []} />
+    /// <Item list={this.props.list ? this.props.list : []} />
     ///
     /// // Good
-    /// <Item callback={this.props.jsx} />
+    /// <Item list={this.props.list} />
     /// ```
-    NoJsxAsProp,
+    NoNewArrayAsProp,
     correctness
 );
 
-impl Rule for NoJsxAsProp {
+impl Rule for NoNewArrayAsProp {
     fn run<'a>(&self, node: &AstNode<'a>, ctx: &LintContext<'a>) {
         if let AstKind::JSXElement(jsx_elem) = node.kind() {
             check_jsx_element(jsx_elem, ctx);
@@ -57,7 +63,7 @@ fn check_jsx_element<'a>(jsx_elem: &JSXElement<'a>, ctx: &LintContext<'a>) {
                 ..
             })) => {
                 if let Some(span) = check_expression(expr) {
-                    ctx.diagnostic(NoJsxAsPropDiagnostic(span));
+                    ctx.diagnostic(NoNewArrayAsPropDiagnostic(span));
                 }
             }
             _ => {}
@@ -67,7 +73,21 @@ fn check_jsx_element<'a>(jsx_elem: &JSXElement<'a>, ctx: &LintContext<'a>) {
 
 fn check_expression(expr: &Expression) -> Option<Span> {
     match expr.without_parenthesized() {
-        Expression::JSXElement(expr) => Some(expr.span),
+        Expression::ArrayExpression(expr) => Some(expr.span),
+        Expression::CallExpression(expr) => {
+            if is_constructor_matching_name(&expr.callee, "Array") {
+                Some(expr.span)
+            } else {
+                None
+            }
+        }
+        Expression::NewExpression(expr) => {
+            if is_constructor_matching_name(&expr.callee, "Array") {
+                Some(expr.span)
+            } else {
+                None
+            }
+        }
         Expression::LogicalExpression(expr) => {
             check_expression(&expr.left).or_else(|| check_expression(&expr.right))
         }
@@ -82,14 +102,18 @@ fn check_expression(expr: &Expression) -> Option<Span> {
 fn test() {
     use crate::tester::Tester;
 
-    let pass = vec![r"<Item callback={this.props.jsx} />"];
+    let pass = vec![r"<Item list={this.props.list} />"];
 
     let fail = vec![
-        r"<Item jsx={<SubItem />} />",
-        r"<Item jsx={this.props.jsx || <SubItem />} />",
-        r"<Item jsx={this.props.jsx ? this.props.jsx : <SubItem />} />",
-        r"<Item jsx={this.props.jsx || (this.props.component ? this.props.component : <SubItem />)} />",
+        r"<Item list={[]} />",
+        r"<Item list={new Array()} />",
+        r"<Item list={Array()} />",
+        r"<Item list={this.props.list || []} />",
+        r"<Item list={this.props.list ? this.props.list : []} />",
+        r"<Item list={this.props.list || (this.props.arr ? this.props.arr : [])} />",
     ];
 
-    Tester::new(NoJsxAsProp::NAME, pass, fail).with_react_perf_plugin(true).test_and_snapshot();
+    Tester::new(NoNewArrayAsProp::NAME, pass, fail)
+        .with_react_perf_plugin(true)
+        .test_and_snapshot();
 }
